refactor: 회원 예약 페이지 url prefix 및 컨트롤러명 수정
