+++ src/main/java/roomescape/admin/AdminPageController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequestMapping("/admin")
-public class AdminController {
+public class AdminPageController {
 
     @GetMapping
     public String index() {
