+++ src/main/java/roomescape/member/MemberPageController.java
@@ -1,13 +1,13 @@
-package roomescape.user;
+package roomescape.member;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class UserController {
+public class MemberPageController {
 
-    @GetMapping("/reservation")
-    public String reservation(){
+    @GetMapping("/member/reservation")
+    public String reservation() {
         return "reservation";
     }
 }
